graph time resolution proper

+++ widget.cpp
@@ -139,20 +139,20 @@ void Widget::draw(View &view, Streams &streams, SDL_Renderer *rend, SDL_Rect &r)
 
 
 #ifndef SAMPLE_FLOAT
-template float  Widget::graph<float >(SDL_Renderer*, SDL_Rect&, float [],           size_t, size_t, float , float , float , float );
-template float  Widget::graph<float >(SDL_Renderer*, SDL_Rect&, float [], float [], size_t, size_t, float , float , float , float );
+template float  Widget::graph<float >(SDL_Renderer*, SDL_Rect&, float [],           size_t, size_t, double , double , float , float );
+template float  Widget::graph<float >(SDL_Renderer*, SDL_Rect&, float [], float [], size_t, size_t, double , double , float , float );
 #endif
 
-template Sample Widget::graph<Sample>(SDL_Renderer*, SDL_Rect&, Sample[],           size_t, size_t, float, float, Sample, Sample);
-template Sample Widget::graph<Sample>(SDL_Renderer*, SDL_Rect&, Sample[], Sample[], size_t, size_t, float, float, Sample, Sample);
+template Sample Widget::graph<Sample>(SDL_Renderer*, SDL_Rect&, Sample[],           size_t, size_t, double, double, Sample, Sample);
+template Sample Widget::graph<Sample>(SDL_Renderer*, SDL_Rect&, Sample[], Sample[], size_t, size_t, double, double, Sample, Sample);
 
-template int    Widget::graph<int   >(SDL_Renderer*, SDL_Rect&, int   [],           size_t, size_t, float, float, int   , int   );
-template int    Widget::graph<int   >(SDL_Renderer*, SDL_Rect&, int   [], int   [], size_t, size_t, float, float, int   , int   );
+template int    Widget::graph<int   >(SDL_Renderer*, SDL_Rect&, int   [],           size_t, size_t, double, double, int   , int   );
+template int    Widget::graph<int   >(SDL_Renderer*, SDL_Rect&, int   [], int   [], size_t, size_t, double, double, int   , int   );
 
 template<typename T>
 T Widget::graph(SDL_Renderer *rend, SDL_Rect &r,
 					 T data[], size_t data_count, size_t stride,
-					 float idx_from, float idx_to,
+					 double idx_from, double idx_to,
 					 T y_min, T y_max)
 {
 	return graph(rend, r, data, data, data_count, stride,
@@ -163,17 +163,17 @@ T Widget::graph(SDL_Renderer *rend, SDL_Rect &r,
 template<typename T>
 T Widget::graph(SDL_Renderer *rend, SDL_Rect &r,
 					 T data_min[], T data_max[], size_t data_count, size_t stride,
-					 float idx_from, float idx_to,
+					 double idx_from, double idx_to,
 					 T y_min, T y_max)
 {
-	float y_scale = (r.h - 2) / ((float)y_min - (float)y_max);
+	float y_scale = (r.h - 2) / ((double)y_min - (float)y_max);
 	float y_off = r.y - (float)y_max * (float)y_scale;
 	T v_peak = 0;
 
 	int npoints = 0;
 	int nrects = 0;
 	int step = 2;
-	float idx_scale = (idx_to - idx_from) / r.w;
+	double idx_scale = (idx_to - idx_from) / r.w;
 	int max_points = r.w / step + 1;
 
 	int idx_min = 0;
