+++ widget.hpp
@@ -47,13 +47,13 @@ protected:
 	template<typename T>
 	T graph(SDL_Renderer *rend, SDL_Rect &r,
 						 T data[], size_t data_count, size_t stride,
-						 float idx_from, float idx_to,
+						 double idx_from, double idx_to,
 						 T y_min, T y_max);
 	
 	template<typename T>
 	T graph(SDL_Renderer *rend, SDL_Rect &r,
 						 T data_min[], T data_max[], size_t data_count, size_t stride,
-						 float idx_from, float idx_to,
+						 double idx_from, double idx_to,
 						 T y_min, T y_max);
 
 	void grid_vertical(SDL_Renderer *rend, SDL_Rect &r, Sample v_from, Sample v_to);
